org.yocto.sdk.docker.launcher: add CROPS-aware container command launcher
This plugin registers YoctoCropsContainerCommandLauncherFactory which has
higher priority than the default factory. This factory only applies to projects
which have the new Yocto Project nature, otherwise it will return null whenever
command launcher for a given project is requested, this allows the
CommandLauncherManager to fallback to lower priority default factory.

YoctoCropsContainerCommandLauncher extends the forked ContainerCommandLauncher
to always not set UID. Note that this has only been tested on CROPS containers.

YoctoCropsContainerLauncher extends the forked ContainerLauncher to reformat
the command and arguments as the container entrypoint's arguments. This
requires YoctoCropsContainerLauncher to have explicit knowledge of the entry
point used by the container and the type of flags supported, as well as how to
remap the working directory, command and arguments for the entry point's flags.

Signed-off-by: Chin Huat Ang <chin.huat.ang@intel.com>