Improve preferences default values and initial values
When a C/C++ project is configured with Yocto Project nature via
EnableNatureAction, an instance of YoctoProjectProjectPreferences will
be created and configured with both default and initial values. This
also means that both project and profile preferences no longer rely on
the editors in YoctoProjectProfileComposedEditor to initialize the
preferences' initial values.

Furthermore, the initial values for the project or profile preferences
are always applied when the preferences are created for the first time,
and this has an added benefit which ensures that any prior preferences
store would also be reinitialized.

Signed-off-by: Chin Huat Ang <chin.huat.ang@intel.com>